+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Added a config option to adjust the transparency of the splashes' white rim
 - Added ripples on water when it's raining
   - Ripple density can be adjusted or disabled in the config
+- Added a config option to control the density or remove flowing water splashing particles
 - Added two new config options to adjust cascade sounds volume and distance
 - Fixed splashes sometimes being completely black (thanks to lonefelidae16)
 - Fixed occasional crashes that could happen randomly (thanks to lonefelidae16)